Code cleanup

+++ GameObject.cpp
@@ -150,7 +150,7 @@ BoundingBox GameObject::getBoundingBox()
 }
 
 //UpdateMatrix() should be called at the end of Update, as it flushes any transformations to the world matrix
-void GameObject::Update(float deltaTime)
+void GameObject::Update(double deltaTime)
 {
 	if (!objectMoves)
 	{
